archive project

+++ src/com/controller/SponsorProjectListController.java
@@ -32,34 +32,44 @@ import com.session.SessionScopeData;
 @Controller
 @RequestMapping("/sponsor/projectList")
 public class SponsorProjectListController {
+	//instantiates UserJdbcServiceImpl for user related activities
+	UserJdbcServiceImpl  sessionService = new UserJdbcServiceImpl();
 	//instantiates ProjectJdbcServiceImpl for user related activities
 	ProjectJdbcServiceImpl  projectService = new ProjectJdbcServiceImpl();
 	//autowire Mail Manager
 	@Autowired
 	private MailManager mailManager;
+	//autowire session variable User
+	@Autowired
+	private SessionScopeData sessionScopeUserData;
 		
 	@RequestMapping(params = {"action", "id", "dispID"})
 	public ModelAndView projectSubmit(@RequestParam(value = "action") String actionVal, @RequestParam(value = "id") int projId, 
 				@RequestParam(value = "dispID", required=false) int dispId, HttpServletRequest request) {
 		ModelAndView mav;
 		String status="";
-		if(actionVal.equals("edit")) {
-			mav = new ModelAndView(new RedirectView("projectForm.do"));
-			mav.addObject("action", "edit");
-			mav.addObject("id", projId);		
-		}
-		else if(actionVal.equals("submit")) {
-			String serverURL = request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort() + request.getContextPath();
-			if(projectService.updateProjectStatusLead(serverURL, mailManager, projId, "Submitted", dispId)>0) {
-				status="project successfully submitted to lead faculty";
+		if(this.sessionCheck(sessionScopeUserData)) {
+			if(actionVal.equals("edit")) {
+				mav = new ModelAndView(new RedirectView("projectForm.do"));
+				mav.addObject("action", "edit");
+				mav.addObject("id", projId);		
+			}
+			else if(actionVal.equals("submit")) {
+				String serverURL = request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort() + request.getContextPath();
+				if(projectService.updateProjectStatusLead(serverURL, mailManager, projId, "Submitted", dispId)>0) {
+					status="project successfully submitted to lead faculty";
+				}
+				else {
+					status="project submission failed";
+				}
+				mav = new ModelAndView(new RedirectView("projectList.do"),"status",status);
 			}
 			else {
-				status="project submission failed";
+				mav = new ModelAndView("sponsor/project");
 			}
-			mav = new ModelAndView(new RedirectView("projectList.do"),"status",status);
 		}
 		else {
-			mav = new ModelAndView("sponsor/project");
+			mav = new ModelAndView(new RedirectView("../login.do"), "status", "Please login first");
 		}
 		return mav;
 	}
@@ -68,36 +78,50 @@ public class SponsorProjectListController {
 	public ModelAndView projectForm(@RequestParam(value = "action") String actionVal, @RequestParam(value = "id") int projId) {
 		ModelAndView mav;
 		String status="";
-		if(actionVal.equals("edit")) {
-			mav = new ModelAndView(new RedirectView("projectForm.do"));
-			mav.addObject("action", "edit");
-			mav.addObject("id", projId);		
-		}
-		else if(actionVal.equals("delete")) {
-			if(projectService.delete(projId)>0) {
-				status="project successfully deleted";
+		if(this.sessionCheck(sessionScopeUserData)) {
+			if(actionVal.equals("edit")) {
+				mav = new ModelAndView(new RedirectView("projectForm.do"));
+				mav.addObject("action", "edit");
+				mav.addObject("id", projId);		
+			}
+			else if(actionVal.equals("delete")) {
+				if(projectService.delete(projId)>0) {
+					status="project successfully deleted";
+				}
+				else {
+					status="project deletion failed";
+				}
+				mav = new ModelAndView(new RedirectView("projectList.do"),"status",status);
+				//mav.addObject("status", status);	
+				
+			}
+			else if(actionVal.equals("archive")) {
+				if(projectService.archiveProject(projId)>0) {
+					status="project successfully archived";
+				}
+				else {
+					status="project archive failed";
+				}
+				mav = new ModelAndView(new RedirectView("archivedProjects.do"),"status",status);
+				//mav.addObject("status", status);	
 			}
 			else {
-				status="project deletion failed";
+				mav = new ModelAndView("sponsor/project");
 			}
-			mav = new ModelAndView(new RedirectView("projectList.do"),"status",status);
-			//mav.addObject("status", status);	
-			
-		}
-		else if(actionVal.equals("archive")) {
-			if(projectService.delete(projId)>0) {
-				status="project successfully archived";
-			}
-			else {
-				status="project deletion failed";
-			}
-			mav = new ModelAndView(new RedirectView("projectList.do"),"status",status);
-			//mav.addObject("status", status);	
 		}
 		else {
-			mav = new ModelAndView("sponsor/project");
+			mav = new ModelAndView(new RedirectView("../login.do"), "status", "Please login first");
 		}
 		return mav;
+	}
+	
+	protected boolean sessionCheck(SessionScopeData sessionData) {		
+		if(sessionScopeUserData!=null) {
+			if(sessionService.sessionCheck(sessionData.getUserInfo().getId())>0) {
+				return true;
+			}				
+		}
+		return false;
 	}
 	
 }
